+++ changelog.md
@@ -1,4 +1,3 @@
 CHECK 3.0.0 PATCH NOTES!!
 
-- fixed sand dupe
-- fixed bamboo offset o planter
+- fixed a bug with trinket compat